Make ticketer tests integration tests.
This is a public API so it should be tested with an integration test.

This also helps with the panic triage.
